Completed all functionality. Need to add final alerts to add/modify customer screens. Need to add javadoc comments and perform final cleanup.

+++ src/controller/ViewAppointmentsScreen.java
@@ -1,18 +1,27 @@
 package controller;
 
 import dao.AppointmentsDao;
+import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.Initializable;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
 import model.Appointment;
+import model.User;
+import utils.alerts.Confirm;
+import utils.alerts.Error;
+import utils.alerts.Login;
+import utils.auth.UserAuth;
 
 import java.io.IOException;
 import java.net.URL;
 import java.sql.SQLException;
 import java.time.DayOfWeek;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 public class ViewAppointmentsScreen implements Initializable {
@@ -27,32 +36,41 @@ public class ViewAppointmentsScreen implements Initializable {
     public RadioButton viewMonthRadioButton;
     public RadioButton viewAllRadioButton;
     public ToggleGroup tGroup;
-    public TableView appointmentsTable;
-    public TableColumn idCol;
-    public TableColumn customerCol;
-    public TableColumn titleCol;
-    public TableColumn descriptionCol;
-    public TableColumn typeCol;
-    public TableColumn locationCol;
-    public TableColumn dateCol;
-    public TableColumn startCol;
-    public TableColumn endCol;
+    public TableView<Appointment> appointmentsTable;
+    public TableColumn<Appointment, Integer> idCol;
+    public TableColumn<Appointment, String> titleCol;
+    public TableColumn<Appointment, String> descriptionCol;
+    public TableColumn<Appointment, String> typeCol;
+    public TableColumn<Appointment, String> locationCol;
+    public TableColumn<Appointment, LocalDateTime> startCol;
+    public TableColumn<Appointment, LocalDateTime> endCol;
+    public TableColumn<Appointment, String> contactCol;
+    public TableColumn<Appointment, Integer> customerIdCol;
+    public TableColumn<Appointment, Integer> userIdCol;
+
+    private static boolean initialLogin = true;
+
+    public static void setInitialLogin(boolean bool) { initialLogin = bool; }
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
         appointmentsTable.setItems(AppointmentsDao.getAllAppointments());
 
         idCol.setCellValueFactory(new PropertyValueFactory<>("id"));
-        customerCol.setCellValueFactory(new PropertyValueFactory<>("customer"));
+        customerIdCol.setCellValueFactory(new PropertyValueFactory<>("customerId"));
         titleCol.setCellValueFactory(new PropertyValueFactory<>("title"));
         descriptionCol.setCellValueFactory(new PropertyValueFactory<>("description"));
         typeCol.setCellValueFactory(new PropertyValueFactory<>("type"));
         locationCol.setCellValueFactory(new PropertyValueFactory<>("location"));
-        dateCol.setCellValueFactory(new PropertyValueFactory<>("date"));
-        startCol.setCellValueFactory(new PropertyValueFactory<>("startTime"));
-        endCol.setCellValueFactory(new PropertyValueFactory<>("endTime"));
+        contactCol.setCellValueFactory(new PropertyValueFactory<>("contact"));
+        startCol.setCellValueFactory(new PropertyValueFactory<>("start"));
+        endCol.setCellValueFactory(new PropertyValueFactory<>("end"));
+        userIdCol.setCellValueFactory(new PropertyValueFactory<>("userId"));
 
         viewAllRadioButton.setSelected(true);
+
+        if (initialLogin)
+            upComingAppointment();
     }
 
     public void onAppointmentsButton(ActionEvent actionEvent) {
@@ -80,29 +98,31 @@ public class ViewAppointmentsScreen implements Initializable {
     }
 
     public void onModifyAppointmentButton(ActionEvent actionEvent) throws IOException {
-        Appointment selectedAppt = (Appointment) appointmentsTable.getSelectionModel().getSelectedItem();
+        Appointment selectedAppt = appointmentsTable.getSelectionModel().getSelectedItem();
         if (selectedAppt == null)
-            // TODO: Error alert
-            System.out.println("No appointment selected.");
-        else
+            Error.noItemSelected();
+        else {
             ModifyAppointmentScreen.setAppointmentToMod(selectedAppt);
             Stage stage = (Stage) modifyAppointmentButton.getScene().getWindow();
             GlobalController.modifyAppointmentScreen(stage);
+        }
     }
 
     public void onDeleteAppointmentButton(ActionEvent actionEvent) throws SQLException {
         Appointment selectedAppt = (Appointment) appointmentsTable.getSelectionModel().getSelectedItem();
         if (selectedAppt == null)
-            // TODO: Error alert
-            System.out.println("No appointment selected.");
+            Error.noItemSelected();
         else {
-            AppointmentsDao.deleteAppointment(selectedAppt);
-            if (viewMonthRadioButton.isSelected())
-                onViewMonthRadioButton(actionEvent);
-            else if (viewWeekRadioButton.isSelected())
-                onViewMonthRadioButton(actionEvent);
-            else
-                onViewAllRadioButton(actionEvent);
+            Optional<ButtonType> userResponse = Confirm.delete();
+            if (userResponse.isPresent() && userResponse.get() == ButtonType.OK) {
+                AppointmentsDao.deleteAppointment(selectedAppt);
+                if (viewMonthRadioButton.isSelected())
+                    onViewMonthRadioButton(actionEvent);
+                else if (viewWeekRadioButton.isSelected())
+                    onViewMonthRadioButton(actionEvent);
+                else
+                    onViewAllRadioButton(actionEvent);
+            }
         }
     }
 
@@ -128,5 +148,23 @@ public class ViewAppointmentsScreen implements Initializable {
 
     public void onViewAllRadioButton(ActionEvent actionEvent) {
         appointmentsTable.setItems(AppointmentsDao.getAllAppointments());
+    }
+
+    public void upComingAppointment() {
+        User user = UserAuth.getLoggedInUser();
+        LocalDateTime now = LocalDateTime.now();
+        boolean apptExists = false;
+        for (Appointment appt : AppointmentsDao.getAllAppointments()) {
+            if (ChronoUnit.MINUTES.between(now, appt.getStart()) <= 15 &&
+                    ChronoUnit.MINUTES.between(now, appt.getStart()) >= 0 &&
+                    appt.getUser().equals(user)) {
+                Login.appointment(appt);
+                apptExists = true;
+            }
+        }
+        if (!apptExists) {
+            Login.appointment();
+        }
+        initialLogin = false;
     }
 }
